src/li-workflows-02-events/requirements.txt
openai==1.55.1
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.40.1
azure-identity==1.19.0
llama-index
llama-index-core==0.12.0
llama-deploy
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
llama-index-program-openai